+++ src/main/java/de/maxhenkel/voicechat/config/ServerConfig.java
@@ -22,8 +22,8 @@ public class ServerConfig {
     public ServerConfig(ConfigBuilder builder) {
         voiceChatPort = builder.integerEntry("port", 24454, 0, 65535);
         voiceChatBindAddress = builder.stringEntry("bind_address", "0.0.0.0");
-        voiceChatDistance = builder.doubleEntry("voice_distance", 32D, 1D, 1_000_000D);
-        voiceChatFadeDistance = builder.doubleEntry("voice_fade_distance", 16D, 1D, 1_000_000D);
+        voiceChatDistance = builder.doubleEntry("max_voice_distance", 48D, 1D, 1_000_000D);
+        voiceChatFadeDistance = builder.doubleEntry("min_voice_distance", 4D, 1D, 1_000_000D);
         crouchDistanceMultiplier = builder.doubleEntry("crouch_distance_multiplier", 1D, 0.01D, 1D);
         voiceChatCodec = builder.enumEntry("codec", Codec.VOIP);
         voiceChatMtuSize = builder.integerEntry("mtu_size", 1024, 256, 10000);